🐞 fix: Resolves error related to missing type declaration, refactors code
