+++ index.ts
@@ -1,2 +1,11 @@
 export { ReactDragDropUpload } from "./src/components/react-drag-drop-upload";
-export type * from "./src/components/react-drag-drop-upload";
+
+export type {
+  ReactDragDropUploadProps,
+  ChildrenUploaderProps,
+  DescriptionProps,
+  DescriptionWrapperProps,
+  HoverMessageProps,
+  DrawTypesProps,
+  UploaderWrapperProps,
+} from "./src/components/react-drag-drop-upload/react-drag-drop-upload";
