+++ src/components/react-drag-drop-upload/index.ts
@@ -1,3 +1,9 @@
-export { default as ReactDragDropUpload } from "./drag-drop-upload";
-export type * from "./drag-drop-upload";
-export type * from "./upload-components";
+export { default as ReactDragDropUpload } from "./react-drag-drop-upload";
+export {
+  Description,
+  DescriptionWrapper,
+  HoverMessage,
+  UploadWrapper,
+  DrawTypes,
+} from "./react-drag-drop-upload";
+export type * from "./react-drag-drop-upload";
