+++ tsconfig.json
@@ -22,6 +22,6 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": ["src", "index.ts"],
+  "include": ["src", "index.ts", "dist"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
